// eshop-web/src/main/resources/client/src/app/article-return/enums/return-order.enum.ts
export enum BATCH_JOB_STATUS {
    HOLD = 'HOLD',
    WAITING = 'WAITING',
    EXECUTING = 'EXECUTING',
    ERROR = 'ERROR',
    FINISHED = 'FINISHED',
    READY = 'READY',
    NOT_RUN = 'NOT_RUN',
    CANCELLING = 'CANCELLING',
    CANCELED = 'CANCELED',
    SCHEDULED = 'SCHEDULED',
}

export enum RETURN_ORDER_MESSAGE_TYPE {
    ERROR = 'ERROR',
    SUCCESS = 'SUCCESS',
    WARNING = 'WARNING'
}